Rename utils.py to lora_utils.py for clarity
- Rename test/utils.py to test/lora_utils.py to avoid conflicts with
  other utils files (e.g., test/srt/cpu/utils.py)
- Update nightly LoRA test imports to use lora_utils
- Compatible with future migration to test/registered
