+++ test/nightly/test_lora_qwen3.py
@@ -15,7 +15,7 @@
 import multiprocessing as mp
 import unittest
 
-from utils import LoRAAdaptor, LoRAModelCase, run_lora_multiple_batch_on_model_cases
+from lora_utils import LoRAAdaptor, LoRAModelCase, run_lora_multiple_batch_on_model_cases
 
 from sglang.test.test_utils import CustomTestCase
